requirements.txt
# ========================================
# 智能口腔健康监测系统 - Python 依赖包
# Python 版本要求：>= 3.10
# ========================================

# ========================================
# Web 框架
# ========================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6  # 文件上传支持

# ========================================
# 数据库
# ========================================
psycopg2-binary==2.9.9  # PostgreSQL 驱动
sqlalchemy==2.0.25  # ORM
asyncpg==0.29.0  # 异步 PostgreSQL 驱动（可选）

# ========================================
# 数据验证
# ========================================
pydantic==2.5.3
pydantic-settings==2.1.0  # 环境变量配置

# ========================================
# 视频/图像处理
# ========================================
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
Pillow==10.2.0  # 图像处理
numpy==1.26.3

# ========================================
# HTTP 客户端（千问 API）
# ========================================
httpx==0.26.0  # 异步 HTTP 客户端
requests==2.31.0  # 同步 HTTP 客户端

# ========================================
# 工具库
# ========================================
python-dotenv==1.0.0  # 环境变量加载
python-jose[cryptography]==3.3.0  # JWT 认证
passlib[bcrypt]==1.7.4  # 密码哈希

# ========================================
# 日志
# ========================================
loguru==0.7.2  # 增强日志库

# ========================================
# 开发工具（可选）
# ========================================
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1  # 代码格式化
flake8==7.0.0  # 代码检查
